refactor: Moves the FromRequestParts impl for the jobs and database to their respective modules
