+++ rapina/src/extract/mod.rs
@@ -91,66 +91,3 @@ macro_rules! impl_deref {
 }
 
 pub(crate) use impl_deref;
-
-// Database extractor (requires "database" feature)
-#[cfg(feature = "database")]
-impl FromRequestParts for crate::database::Db {
-    async fn from_request_parts(
-        _parts: &http::request::Parts,
-        _params: &PathParams,
-        state: &Arc<AppState>,
-    ) -> Result<Self, Error> {
-        use sea_orm::DatabaseConnection;
-
-        let conn = state.get::<DatabaseConnection>().ok_or_else(|| {
-            Error::internal(
-                "Database connection not configured. Did you forget to call .with_database()?",
-            )
-        })?;
-        Ok(crate::database::Db::new(conn.clone()))
-    }
-}
-
-// Jobs extractor (requires "database" feature)
-#[cfg(feature = "database")]
-impl FromRequestParts for crate::jobs::Jobs {
-    async fn from_request_parts(
-        parts: &http::request::Parts,
-        _params: &PathParams,
-        state: &Arc<AppState>,
-    ) -> Result<Self, Error> {
-        use sea_orm::DatabaseConnection;
-
-        let pool = state
-            .get::<DatabaseConnection>()
-            .ok_or_else(|| {
-                Error::internal(
-                    "Database connection not configured. Did you forget to call .with_database()?",
-                )
-            })?
-            .clone();
-
-        let trace_id = parts
-            .extensions
-            .get::<crate::context::RequestContext>()
-            .map(|ctx| ctx.trace_id().to_owned());
-
-        Ok(crate::jobs::Jobs::new(pool, trace_id))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test::{TestRequest, empty_params, empty_state};
-
-    #[cfg(feature = "database")]
-    #[tokio::test]
-    async fn jobs_extractor_missing_db_returns_500() {
-        let (parts, _) = TestRequest::get("/").into_parts();
-        let result =
-            crate::jobs::Jobs::from_request_parts(&parts, &empty_params(), &empty_state()).await;
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err().status(), 500);
-    }
-}
